adding the indivdual event component to go into the calendar gid day component
created the tag that wiould show the title of each event. These tags will go on the date grid that the event falls on. The modal appreing functionality when clicking on these event tags was also added but it is not dynamic yet (does not adjust for each events unique data).