Got the entire grid days container to be responsive to tag overflows
Got the grid days container to now always display the same max amount of event tags. Before this, every time the individual grid day cells rendered the event tags, the individual cells would render in order, causing the cells at the start to take up enough space for the max amount of events, which subsequently distructed the cell height calculation for cells in remaining rows. By moving the calculation to the entire days grid container, the height of each cell is calculated before the cell and the max possible amount of event tags in that cell is calculated before the individual day grid cell is rendered and the max number of event tags is passed into the cell, which then renders up to a maximum of that many event tags.